requirements.txt
streamlit==1.28.1
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2
